Module-4_AI_meeting_companion/requirements.txt
groq>=0.4.0
gradio>=4.0.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0